+++ src/test/suite/versionUtils.test.ts
@@ -106,20 +106,20 @@ suite("VersionUtils check", () => {
 			await validateVersion('2.2.0', Platform.LINUX, 'https://github.com/apache/camel-k/releases/download/v2.2.0/camel-k-client-2.2.0-linux-amd64.tar.gz');
 		});
 
-		test("validate url for existing 2.3.0 version", async () => {
-			await validateVersion('2.3.0', Platform.LINUX, 'https://github.com/apache/camel-k/releases/download/v2.3.0/camel-k-client-2.3.0-linux-amd64.tar.gz');
-		});
-
 		test("validate url for existing 2.3.1 version", async () => {
 			await validateVersion('2.3.1', Platform.LINUX, 'https://github.com/apache/camel-k/releases/download/v2.3.1/camel-k-client-2.3.1-linux-amd64.tar.gz');
 		});
 
-		test("validate url for existing 2.3.1 windows version", async () => {
-			await validateVersion('2.3.1', Platform.WINDOWS, 'https://github.com/apache/camel-k/releases/download/v2.3.1/camel-k-client-2.3.1-windows-amd64.tar.gz');
+		test("validate url for existing 2.3.2 version", async () => {
+			await validateVersion('2.3.2', Platform.LINUX, 'https://github.com/apache/camel-k/releases/download/v2.3.2/camel-k-client-2.3.2-linux-amd64.tar.gz');
 		});
 
-		test("validate url for existing 2.3.1 MacOS version", async () => {
-			await validateVersion('2.3.1', Platform.MACOS, 'https://github.com/apache/camel-k/releases/download/v2.3.1/camel-k-client-2.3.1-darwin-amd64.tar.gz');
+		test("validate url for existing 2.3.2 windows version", async () => {
+			await validateVersion('2.3.2', Platform.WINDOWS, 'https://github.com/apache/camel-k/releases/download/v2.3.2/camel-k-client-2.3.2-windows-amd64.tar.gz');
+		});
+
+		test("validate url for existing 2.3.2 MacOS version", async () => {
+			await validateVersion('2.3.2', Platform.MACOS, 'https://github.com/apache/camel-k/releases/download/v2.3.2/camel-k-client-2.3.2-darwin-amd64.tar.gz');
 		});
 
 		test("validate invalid url for xyz1 version", async () => {