+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "vscode-camelk" extension will be documented in this 
 
 ## 0.0.42
 
+- Update default runtime version to v2.3.2
+
 ## 0.0.41
 
 - Update Red Hat telemetry dependency to reduce number of event sent (startup information will be sent one time per day per user)